change images - delete past days response from available times - checking upcomming/past

+++ database/migrations/2026_04_26_165128_create_growths_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('child_id')->constrained('children')->cascadeOnDelete();
             $table->float('height');
-            $table->integer('weight');
+            $table->float('weight');
             $table->date('date');
             $table->timestamps();
         });
